\name{afdiff}
\alias{afdiff}
\title{afdiff}
\usage{
  afdiff(listOfFiles = NULL,
    ComputeBackwardDifference = FALSE,
    ComputeCentralDifference = FALSE, Channel = 1,
    ToFile = TRUE, ExplicitExt = NULL)
}
\arguments{
  \item{listOfFiles}{vector of file paths to be processed
  by function}

  \item{ComputeBackwardDifference}{}

  \item{ComputeCentralDifference}{}

  \item{Channel}{}

  \item{ToFile}{}

  \item{ExplicitExt}{}
}
\value{
  nr of files processed
}
\description{
  afdiff adapted from assp
}
\details{
  still have to write propper manual entry
}
\author{
  Raphael Winkelmann
}

