\name{affilter}
\alias{affilter}
\title{affilter}
\usage{
  affilter(listOfFiles = NULL, HighPass = 4000,
    LowPass = 0, StopBand = 96, Transition = 250,
    UseIIR = FALSE, NumIIRsections = 4, ToFile = TRUE,
    ExplicitExt = NULL)
}
\arguments{
  \item{listOfFiles}{vector of file paths to be processed
  by function}

  \item{HighPass}{bli}

  \item{LowPass}{blup}

  \item{StopBand}{bla}

  \item{Transition}{bli}

  \item{UseIIR}{blup}

  \item{NumIIRsections}{bla}

  \item{ToFile}{bla}

  \item{ExplicitExt}{bli}
}
\value{
  number of files processed
}
\description{
  affilter function adapted from assp
}
\details{
  still have to write propper man pages
}
\author{
  Raphael Winkelmann
}

