% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspTypes.R
\name{AsspLpTypes}
\alias{AsspLpTypes}
\title{AsspLpTypes}
\usage{
AsspLpTypes()
}
\value{
vector containing lp types
}
\description{
returns all valid AsspLpTypes according to the assp library
}
\details{
wrapper function for AsspLpTypes of wrassp
}
\author{
Raphael Winkelmann
}

