\name{AsspWindowTypes}
\alias{AsspWindowTypes}
\title{AsspWindowTypes}
\usage{
  AsspWindowTypes()
}
\value{
  vector containing window types
}
\description{
  returns all valid AsspWindowTypes according to the assp
  library
}
\details{
  wrapper function for AsspWindowTypes of wrassp
}
\author{
  Raphael Winkelmann
}

