\docType{data}
\name{forceToLogDefault}
\alias{forceToLogDefault}
\title{forceToLogDefault (BOOL)}
\format{logi TRUE}
\usage{
  forceToLogDefault
}
\description{
  Bool switch as default for logging options
}
\details{
  this is used to switch of force logging of all signal
  processing functions
}
\author{
  Raphael Winkelmann
}
\keyword{datasets}

