\name{getDObj}
\alias{getDObj}
\title{getDObj}
\usage{
  getDObj(fname, begin = 0, end = 0, samples = FALSE)
}
\arguments{
  \item{fname}{file name to retrieve}

  \item{begin}{begin time (default is in ms) of segment to
  retrieve}

  \item{end}{end time (default is in ms) of segment to
  retrieve}

  \item{samples}{(BOOL) if set to false ms values of
  begin/end are sample numbers}
}
\value{
  list object containing file data
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

