\name{hasDuplicateFiles}
\alias{hasDuplicateFiles}
\title{hasDupicateFiles}
\usage{
  hasDuplicateFiles(listOfFilePaths, newExt)
}
\arguments{
  \item{listOfFilePaths}{}

  \item{newExt}{}
}
\value{
  (BOOL)
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

