\name{is.AsspDataObj}
\alias{is.AsspDataObj}
\title{Checks whether x is a valid AsspDataObj}
\usage{
  is.AsspDataObj(x, ...)
}
\arguments{
  \item{x}{an object of class AsspDataObj}

  \item{...}{optional other arguments passed to further
  functions}
}
\value{
  TRUE or FALSE
}
\description{
  .. content for description{} (no empty lines) ..
}
\details{
  .. content for details{} ..
}
\author{
  Lasse Bombien
}

