\name{isAsspLpType}
\alias{isAsspLpType}
\title{isAsspLpType}
\usage{
  isAsspLpType(lpName = NULL)
}
\arguments{
  \item{lpName}{name of lp type}
}
\value{
  (BOOL) true if lpName is valid; false otherwise
}
\description{
  checks if given string is a valid AsspLpType according to
  the assp library
}
\author{
  Raphael Winkelmann
}

