\name{isAsspWindowType}
\alias{isAsspWindowType}
\title{isAsspWindowType}
\usage{
  isAsspWindowType(windowName = NULL)
}
\arguments{
  \item{windowName}{name of window}
}
\value{
  (BOOL) true if windowName is valid
}
\description{
  this is the description
}
\details{
  and these are the details
}
\author{
  Raphael Winkelmann
}

