\name{print.AsspDataObj}
\alias{print.AsspDataObj}
\title{print.AsspDataObj}
\usage{
  print.AsspDataObj(dobj)
}
\arguments{
  \item{dobj}{an object of class AsspDataObj}
}
\description{
  Prints an overview of ASSP Data Objects
}
\details{
  nothing really
}
\author{
  Lasse Bombien
}
\seealso{
  \code{\link{getDObj}}
}

