\docType{data}
\name{useWrasspLogger}
\alias{useWrasspLogger}
\title{package variable to force the usage of the logger
set to FALSE by defaul}
\format{logi FALSE}
\usage{
  useWrasspLogger
}
\description{
  package variable to force the usage of the logger set to
  FALSE by defaul
}
\author{
  Raphael Winkelmann
}
\keyword{datasets}

