\docType{data}
\name{wrasspOutputInfos}
\alias{wrasspOutputInfos}
\title{list of default output extensions,
track names and output type
for each signal processing function in wrassp}
\format{List of 13
 $ acfana     :List of 3
 $ afdiff     :List of 3
 $ affilter   :List of 3
 $ cepstrum   :List of 3
 $ cssSpectrum:List of 3
 $ dftSpectrum:List of 3
 $ f0_ksv     :List of 3
 $ f0_mhs     :List of 3
 $ forest     :List of 3
 $ lpsSpectrum:List of 3
 $ rfcana     :List of 3
 $ rmsana     :List of 3
 $ zcrana     :List of 3}
\usage{
  wrasspOutputInfos
}
\description{
  list of default output extensions, track names and output
  type for each signal processing function in wrassp
}
\author{
  Raphael Winkelmann
}
\keyword{datasets}

