% Generated by roxygen2 (4.0.1): do not edit by hand
\name{affilter}
\alias{affilter}
\title{affilter}
\usage{
affilter(listOfFiles = NULL, optLogFilePath = NULL, HighPass = 4000,
  LowPass = 0, StopBand = 96, Transition = 250, UseIIR = FALSE,
  NumIIRsections = 4, ToFile = TRUE, ExplicitExt = NULL,
  OutputDirectory = NULL, forceToLog = useWrasspLogger)
}
\arguments{
\item{listOfFiles}{vector of file paths to be processed by function}

\item{optLogFilePath}{path to option log file}

\item{HighPass}{= <num>: set the high-pass cut-off frequency to <num> Hz (default: 0, no high-pass filtering)}

\item{LowPass}{= <num>: set the low-pass cut-off frequency to <num> Hz (default: 0, no low-pass filtering)}

\item{StopBand}{= <num>: set the stop-band attenuation to <num> dB (default: 93.0 dB, minimum: 21.0 dB)}

\item{Transition}{= <num>: set the width of the transition band to <num> Hz (default: 250.0 Hz)}

\item{UseIIR}{switch from the default FIR to IIR filter}

\item{NumIIRsections}{= <num>: set the number of 2nd order sections to <num> (default: 4) where each section adds 12dB/oct to the slope of the filter}

\item{ToFile}{write results to file (for default extension see details section))}

\item{ExplicitExt}{set if you wish to overwride the default extension}

\item{OutputDirectory}{directory in which output files are stored. Defaults to NULL, i.e.
the directory of the input files}

\item{forceToLog}{is set by the global package variable useWrasspLogger. This is set
to FALSE by default and should be set to TRUE is logging is desired.}
}
\value{
nrOfProcessedFiles or if only one file to process return AsspDataObj of that file
}
\description{
affilter function adapted from libassp
}
\details{
Filters the audio signal in <listOfFiles>.
By specifying the high-pass and/or low-pass cut-off
frequency one of four filter characteristics may be
selected as shown in the table below.\cr
hp     lp     filter characteristic         extension\cr
-------------------------------------------------------\cr
> 0     [0]    high-pass from hp             '.hpf'\cr
 [0]   > 0     low-pass up to lp             '.lpf'\cr
> 0    > hp    band-pass from hp to lp       '.bpf'\cr
> lp   > 0     band-stop between lp and hp   '.bsf'\cr
The Kaiser-window design method is used to compute the
coefficients of a linear-phase FIR filter with unity gain
in the pass-band. The cut-off frequencies (-6 dB points)
of the filters are in the middle of the transition band.
The filtered signal will be written to a file with the
base name of the input file and an extension corresponding
to the filter characteristic (see table). The format of
the output file will be the same as that of the input file.
}
\author{
Raphael Winkelmann
}

