% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AsspLpTypes}
\alias{AsspLpTypes}
\title{AsspLpTypes}
\usage{
AsspLpTypes()
}
\value{
vector containing lp types
}
\description{
returns all valid AsspLpTypes according to the assp library
}
\details{
wrapper function for AsspLpTypes of wrassp
}
\author{
Raphael Winkelmann
}

