% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspTypes.R
\name{AsspSpectTypes}
\alias{AsspSpectTypes}
\title{AsspSpectTypes}
\usage{
AsspSpectTypes()
}
\value{
vector containing spectrogram types
}
\description{
returns all valid AsspSpectTypes according to the assp library
}
\details{
wrapper function for AsspSpectTypes of wrassp
}
\author{
Raphael Winkelmann
}
