% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delTrack}
\alias{delTrack}
\title{Remove track from an AsspDataObj}
\usage{
delTrack(dobj, trackname)
}
\arguments{
\item{dobj}{An object of class AsspDataObj}

\item{trackname}{the name of a track in this object}
}
\value{
The object without the track named trackname
}
\description{
Remove a track from an
AsspDataObj object
}
\author{
Lasse Bombien
}

