\name{f0ana}
\alias{f0ana}
\title{f0ana}
\usage{
  f0ana(listOfFiles = NULL, BeginTime = 0, EndTime = 0,
    WindowShift = 5, Gender = "u", MaxF = 600, MinF = 50,
    MinAmp = 50, MaxZCR = 3000, ToFile = TRUE,
    ExplicitExt = NULL)
}
\arguments{
  \item{listOfFiles}{vector of file paths to be processed
  by function}

  \item{BeginTime}{bla}

  \item{EndTime}{bli}

  \item{WindowShift}{blup}

  \item{Gender}{bla}

  \item{MaxF}{bli}

  \item{MinF}{blup}

  \item{MinAmp}{bla}

  \item{MaxZCR}{bli}

  \item{ToFile}{write results to file (default extension is
  .f0)}

  \item{ExplicitExt}{set if you wish to overwride the
  default extension}
}
\value{
  nrOfProcessedFiles or if only one file to process return
  dataObj of that file
}
\description{
  this is the description
}
\details{
  and these are the details
}
\author{
  Raphael Winkelmann
}

