% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspDataObj.R
\name{is.AsspDataObj}
\alias{is.AsspDataObj}
\title{Checks whether x is a valid AsspDataObj}
\usage{
is.AsspDataObj(x, ...)
}
\arguments{
\item{x}{an object of class AsspDataObj}

\item{...}{optional other arguments passed to further functions}
}
\value{
TRUE or FALSE
}
\description{
Checks whether x is a valid AsspDataObj
}
\author{
Lasse Bombien
}
