% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.AsspDataObj}
\alias{is.AsspDataObj}
\title{Checks whether x is a valid AsspDataObj}
\usage{
is.AsspDataObj(x, ...)
}
\arguments{
\item{x}{an object of class AsspDataObj}

\item{...}{optional other arguments passed to further functions}
}
\value{
TRUE or FALSE
}
\description{
Checks whether x is a valid AsspDataObj
}
\author{
Lasse Bombien
}

