\name{isAsspLpType}
\alias{isAsspLpType}
\title{isAsspLpType}
\usage{
  isAsspLpType(lpName = NULL)
}
\arguments{
  \item{lpName}{}
}
\value{
  (BOOL) true if lpName is valid
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

