\name{isAsspLpType}
\alias{isAsspLpType}
\title{isAsspLpType}
\usage{
  isAsspLpType(lpName = NULL)
}
\arguments{
  \item{lpName}{name of lp type}
}
\value{
  (BOOL) true if lpName is valid
}
\description{
  this is the description
}
\details{
  and these are the details
}
\author{
  Raphael Winkelmann
}

