% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspChecks.R
\name{isAsspSpectType}
\alias{isAsspSpectType}
\title{isAsspSpectType}
\usage{
isAsspSpectType(spectName = NULL)
}
\arguments{
\item{spectName}{name of lp type}
}
\value{
(BOOL) true if spectName is valid; false otherwise
}
\description{
checks if given string is a valid AsspSpectType according to the assp library
}
\author{
Raphael Winkelmann
}

