\name{isAsspWindowType}
\alias{isAsspWindowType}
\title{isAsspWindowType}
\usage{
  isAsspWindowType(windowName = NULL)
}
\arguments{
  \item{windowName}{}
}
\value{
  (BOOL) true if windowName is valid
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

