% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isAsspWindowType}
\alias{isAsspWindowType}
\title{isAsspWindowType}
\usage{
isAsspWindowType(windowName = NULL)
}
\arguments{
\item{windowName}{name of window}
}
\value{
(BOOL) true if windowName is valid; false otherwise
}
\description{
checks if given string is a valid AsspWindowType according to the assp library
}
\author{
Raphael Winkelmann
}

