% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspChecks.R
\name{isAsspWindowType}
\alias{isAsspWindowType}
\title{isAsspWindowType}
\usage{
isAsspWindowType(windowName = NULL)
}
\arguments{
\item{windowName}{name of window}
}
\value{
(BOOL) true if windowName is valid; false otherwise
}
\description{
checks if given string is a valid AsspWindowType according to the assp library
}
\author{
Raphael Winkelmann
}
