\name{mhspitch}
\alias{mhspitch}
\title{mhspitch}
\usage{
  mhspitch(listOfFiles = NULL, BeginTime = 0,
    CenterTime = FALSE, EndTime = 0, WindowShift = 5,
    Gender = "u", MaxF = 600, MinF = 50, MinAmp = 50,
    MinAC1 = 0.25, MinRMS = 18, MaxZCR = 3000,
    MinProb = 0.52, PlainSpectrum = FALSE, ToFile = TRUE,
    ExplicitExt = NULL)
}
\arguments{
  \item{listOfFiles}{vector of file paths to be processed
  by function}

  \item{BeginTime}{bla}

  \item{CenterTime}{bli}

  \item{EndTime}{blup}

  \item{WindowShift}{bla}

  \item{Gender}{bli}

  \item{MaxF}{blup}

  \item{MinF}{bla}

  \item{MinAmp}{bli}

  \item{MinAC1}{blup}

  \item{MinRMS}{bla}

  \item{MaxZCR}{bli}

  \item{MinProb}{blup}

  \item{PlainSpectrum}{bla}

  \item{ToFile}{bli}

  \item{ExplicitExt}{blup}
}
\value{
  nrOfProcessedFiles or if only one file to process return
  dataObj of that file
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

