% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHelper.R
\name{prepareFiles}
\alias{prepareFiles}
\title{Normalise a list of filenames so that they can be passed to a signal processing function}
\usage{
prepareFiles(listOfFiles)
}
\arguments{
\item{listOfFiles}{The list of file names to process}
}
\value{
A normalised list of filenames
}
\description{
Normalise a list of filenames so that they can be passed to a signal processing function
}
\examples{


#   listOfFiles <- prepareFiles(listOfFiles)


}
\author{
Raphael Winkelmann
}
