\name{print.AsspDataObj}
\alias{print.AsspDataObj}
\title{print aa summary of an AsspDataObj}
\usage{
  \method{print}{AsspDataObj} (x, ...)
}
\arguments{
  \item{x}{an object of class AsspDataObj}

  \item{...}{other arguments that might be passed on to
  other functions}
}
\description{
  Prints an overview of ASSP Data Objects
}
\details{
  nothing really
}
\author{
  Lasse Bombien
}
\seealso{
  \code{\link{read.AsspDataObj}}
}

