% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsspDataObj.R
\name{print.AsspDataObj}
\alias{print.AsspDataObj}
\alias{summary.AsspDataObj}
\title{print a summary of an AsspDataObj}
\usage{
\method{print}{AsspDataObj}(x, ...)
}
\arguments{
\item{x}{an object of class AsspDataObj}

\item{...}{other arguments that might be passed on to other functions}
}
\description{
Prints an overview of ASSP Data Objects
}
\author{
Lasse Bombien
}
\seealso{
\code{\link{read.AsspDataObj}}
}

