\name{rmsana}
\alias{rmsana}
\title{rmsana}
\usage{
  rmsana(listOfFiles = NULL, BeginTime = 0,
    CenterTime = FALSE, EndTime = 0, WindowShift = 5,
    WindowSize = 20, EffectiveLength = TRUE,
    Linear = FALSE, Window = "HAMMING", ToFile = TRUE,
    ExplicitExt = NULL)
}
\arguments{
  \item{listOfFiles}{nrOfProcessedFiles or if only one file
  to process return dataObj of that file}

  \item{BeginTime}{bla}

  \item{CenterTime}{bli}

  \item{EndTime}{blup}

  \item{WindowShift}{bla}

  \item{WindowSize}{bli}

  \item{EffectiveLength}{blup}

  \item{Linear}{bla}

  \item{Window}{bli}

  \item{ToFile}{blup}

  \item{ExplicitExt}{}
}
\value{
  nrOfProcessedFiles or if only one file to process return
  dataObj of that file
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Raphael Winkelmann
}

