% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packageVars.R
\docType{data}
\name{useWrasspLogger}
\alias{useWrasspLogger}
\title{package variable to force the usage of the logger
set to FALSE by default}
\format{\preformatted{ logi FALSE
}}
\usage{
useWrasspLogger
}
\description{
package variable to force the usage of the logger
set to FALSE by default
}
\author{
Raphael Winkelmann
}
\keyword{datasets}

