% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{wrasspOutputInfos}
\alias{wrasspOutputInfos}
\title{list of default output extensions,
track names and output type
for each signal processing function in wrassp}
\format{\preformatted{List of 13
 $ acfana     :List of 3
  ..$ ext       : chr "acf"
  ..$ tracks    : chr "acf"
  ..$ outputType: chr "SSFF"
 $ afdiff     :List of 3
  ..$ ext       : chr "dwav"
  ..$ tracks    : chr ""
  ..$ outputType: chr "wav"
 $ affilter   :List of 3
  ..$ ext       : chr [1:4] "hpf" "lpf" "bpf" "bsf"
  ..$ tracks    : chr ""
  ..$ outputType: chr "wav"
 $ cepstrum   :List of 3
  ..$ ext       : chr "cep"
  ..$ tracks    : chr "cep"
  ..$ outputType: chr "SSFF"
 $ cssSpectrum:List of 3
  ..$ ext       : chr "css"
  ..$ tracks    : chr "css"
  ..$ outputType: chr "SSFF"
 $ dftSpectrum:List of 3
  ..$ ext       : chr "dft"
  ..$ tracks    : chr "dft"
  ..$ outputType: chr "SSFF"
 $ ksvF0      :List of 3
  ..$ ext       : chr "f0"
  ..$ tracks    : chr "F0"
  ..$ outputType: chr "SSFF"
 $ mhsF0      :List of 3
  ..$ ext       : chr "pit"
  ..$ tracks    : chr "pitch"
  ..$ outputType: chr "SSFF"
 $ forest     :List of 3
  ..$ ext       : chr "fms"
  ..$ tracks    : chr [1:2] "fm" "bw"
  ..$ outputType: chr "SSFF"
 $ lpsSpectrum:List of 3
  ..$ ext       : chr "lps"
  ..$ tracks    : chr "lps"
  ..$ outputType: chr "SSFF"
 $ rfcana     :List of 3
  ..$ ext       : chr [1:4] "rfc" "arf" "lar" "lpc"
  ..$ tracks    : chr [1:3] "rms" "gain" "arf|lar|lpc|rfc"
  ..$ outputType: chr "SSFF"
 $ rmsana     :List of 3
  ..$ ext       : chr "rms"
  ..$ tracks    : chr "rms"
  ..$ outputType: chr "SSFF"
 $ zcrana     :List of 3
  ..$ ext       : chr "zcr"
  ..$ tracks    : chr "zcr"
  ..$ outputType: chr "SSFF"
}}
\usage{
wrasspOutputInfos
}
\description{
list of default output extensions,
track names and output type
for each signal processing function in wrassp
}
\author{
Raphael Winkelmann
}
\keyword{datasets}

