% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageVars.R
\docType{data}
\name{wrasspOutputInfos}
\alias{wrasspOutputInfos}
\title{list of default output extensions,
track names and output type 
for each signal processing function in wrassp}
\format{An object of class \code{list} of length 13.}
\usage{
wrasspOutputInfos
}
\description{
list of default output extensions,
track names and output type 
for each signal processing function in wrassp
}
\author{
Raphael Winkelmann
}
\keyword{datasets}
